+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - BREAKING: Public API change.  The `bundle()` method now takes a manifest
   instead of entrypoints, strategy and mapper.  To produce a manifest,
   use new public method `generateManifest()`.
+- Fixed an issue where an immediate `<style>` child of `<dom-module>` was
+  not moved into generated `<template>`.
 
 ## 2.0.0-pre.11 - 2017-03-20
 - Bump dependency on analyzer